requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.12
pymysql==1.0.3
cryptography==40.0.2
pydantic==1.10.7
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator==2.0.0
alembic==1.10.4
httpx==0.24.0
authlib==1.2.0
itsdangerous==2.1.2
openai==0.27.6
python-dotenv==1.0.0

